Remove old summarization modules
