--- src/datasets/summarizations/utils/constants.ts
@@ -1,2 +1,0 @@
-// The size of the chart is fixed to 800 * 500
-export const CHART_DIAGONAL_ANGLE = Math.atan(500 / 800);
