+++ src/datasets/user-workday-holiday-linear.dataset.ts
@@ -6,10 +6,7 @@ import { createTimeSeriesQuery } from './queries/time-series.query';
 import { createLineChartMeta } from './metas/line-chart.meta';
 import { PreferenceMeta } from '../services/preference/types';
 import { DAY } from '../utils/timeUnits';
-import { combineQuerySummariesFactories } from './summarizations/utils/commons';
-import { groupPointsByXWeek } from './summarizations/utils/time-series';
 import { SummarizationMeta, SUMMARIZATION } from '../services/summarization/types';
-import * as TrendSummarization from './summarizations/trend-regression.summarization';
 
 export interface Config {
   dailyWeightStd: number;
@@ -80,10 +77,6 @@ export function create(config: Config): Dataset {
   };
 
   const dataCube = generateCube(categories, measures, generateCubeConfig);
-
-  const activeUserQuerySummariesFactory = combineQuerySummariesFactories(
-    TrendSummarization.queryFactory,
-  );
 
   const lineChartMeta = createLineChartMeta(
     'Active Users',
